move the webserver into a web module
