+++ src/web/mod.rs
@@ -1,0 +1,3 @@
+mod router;
+
+pub use self::router::Router;
